pyproject.toml
[project]
name = "sd-mcp-server"
version = "0.1.0"
description = "Stable Diffusion MCP Server for LM Studio integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "discord-py>=2.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "nltk>=3.8.0",
]

[tool.uv]
dev-dependencies = [
    "pillow>=10.0.0",  # GUI testing tool
]
